refactor: 변수 위치 수정
