--- server/src/constants/clientUrl.ts
@@ -1,3 +1,0 @@
-const CLIENT_URL = process.env.NODE_ENV === "production" ? "https://byjuun.com" : "https://local.byjuun.com:3000";
-
-export default CLIENT_URL;
